issue2 - Nonowner edit/delete image
Fixed the bug - Only the owner of the image can edit/delete the image and the app will throw appropriate error message for others

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,22 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
 
-        String tags = convertTagsToString(image.getTags());
+        String error = "Only the owner of the image can edit the image";
+        User user = (User) session.getAttribute("loggeduser");
+        //String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        model.addAttribute("tags", image.getTags());
+
+        String imageUsername = image.getUser().getUsername();
+
+        if (!imageUsername.equals(user.getUsername())) {
+            model.addAttribute("editError", error);
+            return "images/image";
+        } else
+             return "images/edit";
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +149,23 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+        Image image = imageService.getImage(imageId);
+
+        String error = "Only the owner of the image can delete the image";
+        User user = (User) session.getAttribute("loggeduser");
+        model.addAttribute("image", image);
+        model.addAttribute("tags", image.getTags());
+
+        String imageUsername = image.getUser().getUsername();
+
+        if (!imageUsername.equals(user.getUsername())) {
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        } else {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
     }
